Code Review Fixes
