+++ TMS.BusinessGateway/Task/Order.cs
@@ -1,4 +1,6 @@
-﻿using Newtonsoft.Json;
+﻿using ActiveUp.Net.Mail;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
 using RestSharp;
 using System;
 using System.Collections.Generic;
@@ -8,6 +10,7 @@ using System.Linq;
 using System.Net;
 using System.Text;
 using System.Threading.Tasks;
+using TMS.BusinessGateway.Classes;
 using TMS.DataGateway.Repositories.Interfaces;
 using TMS.DomainGateway.Task;
 using TMS.DomainObjects.Objects;
@@ -30,6 +33,22 @@ namespace TMS.BusinessGateway.Task
             if (requestQueryParameter != null)
             {
                 request.AddJsonBody(requestQueryParameter);
+            }
+            var result = client.Execute(request);
+            return result.Content;
+        }
+
+        private static string GetFileUploadApiResponse(string apiRoute, Method method, MimePart mimePart, string token)
+        {
+            var client = new RestClient(ConfigurationManager.AppSettings["ApiGatewayBaseURL"]);
+            if (token != null)
+                client.AddDefaultHeader("Token", token);
+            var request = new RestRequest(apiRoute, method) { RequestFormat = DataFormat.Json, AlwaysMultipartFormData = true };
+            request.Timeout = 500000;
+            if (mimePart != null)
+            {
+                client.AddDefaultHeader("Content-Type", "multipart/form-data");
+                request.AddFileBytes("file", mimePart.BinaryContent, mimePart.Filename, mimePart.ContentType.MimeType);// upload from file byte array
             }
             var result = client.Execute(request);
             return result.Content;
@@ -670,7 +689,43 @@ namespace TMS.BusinessGateway.Task
 
         public override PackingSheetResponse CreateUpdatePackingSheet(PackingSheetRequest packingSheetRequest)
         {
-            PackingSheetResponse  packingSheetResponse = _orderRepository.CreateUpdatePackingSheet(packingSheetRequest);
+            PackingSheetResponse packingSheetResponse = _orderRepository.CreateUpdatePackingSheet(packingSheetRequest);
+
+            if (packingSheetResponse.Status == DomainObjects.Resource.ResourceData.Success && packingSheetResponse.StatusCode == (int)HttpStatusCode.OK && packingSheetResponse.Data.Count > 0)
+            {
+
+                foreach (PackingSheet ps in packingSheetRequest.Requests)
+                {
+                    ps.OrderNumber = packingSheetResponse.Data[0].OrderNumber;
+
+                    ps.DealerNumber = (from pks in packingSheetResponse.Data
+                                       where pks.DealerId == ps.DealerId
+                                       select pks.DealerNumber).FirstOrDefault();
+                    ps.DealerId = 0;
+                    ps.OrderDetailId = 0;
+                }
+
+                #region Call OMS API to update packing sheet details
+
+                LoginRequest omsLoginRequest = new LoginRequest();
+                string omsToken = "";
+                omsLoginRequest.UserName = ConfigurationManager.AppSettings["OMSLogin"];
+                omsLoginRequest.UserPassword = ConfigurationManager.AppSettings["OMSPassword"];
+                var tmsLoginResponse = JsonConvert.DeserializeObject<UserResponse>(GetApiResponse(ConfigurationManager.AppSettings["ApiGatewayOMSURL"]
+                                                                                                  + "v1/user/login", Method.POST, omsLoginRequest, null));
+                if (tmsLoginResponse != null && tmsLoginResponse.Data.Count > 0)
+                {
+                    omsToken = tmsLoginResponse.TokenKey;
+                }
+
+                PackingSheetResponse omsPackingSheetResponse = JsonConvert.DeserializeObject<PackingSheetResponse>(GetApiResponse(ConfigurationManager.AppSettings["ApiGatewayOMSURL"]
+                                                                                                          + "v1/order/CreateUpdatePackingSheet", Method.POST, packingSheetRequest, omsToken));
+
+                packingSheetResponse.StatusMessage = omsPackingSheetResponse.StatusMessage + " " + packingSheetResponse.StatusMessage;
+
+                #endregion
+            }
+
             return packingSheetResponse;
         }
 
@@ -694,7 +749,7 @@ namespace TMS.BusinessGateway.Task
         {
             //If needed write business logic here for request.
 
-            CommonResponse commonResponse  = _orderRepository.GetOrderIds(tokenValue);
+            CommonResponse commonResponse = _orderRepository.GetOrderIds(tokenValue);
 
             //If needed write business logic here for response.
             return commonResponse;
@@ -724,12 +779,170 @@ namespace TMS.BusinessGateway.Task
 
         public override OrderStatusResponse UpdateOrderStatus(OrderStatusRequest request)
         {
-            return _orderRepository.UpdateOrderStatus(request);
+            OrderStatusResponse response = _orderRepository.UpdateOrderStatus(request);
+
+            #region Update Status to OMS 
+            OrderStatusRequest omsRequest = new OrderStatusRequest()
+            {
+                Requests = new List<OrderStatus>()
+            };
+
+            foreach (var item in request.Requests)
+            {
+                OrderStatus requestData = new OrderStatus()
+                {
+                    IsLoad = null,
+                    OrderNumber = item.OrderNumber,
+                    OrderStatusCode = item.OrderStatusCode,
+                    Remarks = "",
+                    SequenceNumber = item.SequenceNumber,
+                    NewSequenceNumber = item.NewSequenceNumber
+                };
+
+                omsRequest.Requests.Add(requestData);
+                omsRequest.RequestFrom = "TMS";
+            }
+
+            if (omsRequest.Requests.Count > 0)
+            {
+                #region Call OMS API to Update Order
+                #region Login to OMS and get Token
+                LoginRequest loginRequest = new LoginRequest();
+                string token = "";
+
+                loginRequest.UserName = ConfigurationManager.AppSettings["OMSLogin"];
+                loginRequest.UserPassword = ConfigurationManager.AppSettings["OMSPassword"];
+                var tmsLoginResponse = JsonConvert.DeserializeObject<UserResponse>(GetApiResponse(ConfigurationManager.AppSettings["ApiGatewayOMSURL"]
+                    + "/v1/user/login", Method.POST, loginRequest, null));
+                if (tmsLoginResponse != null && tmsLoginResponse.Data.Count > 0)
+                {
+                    token = tmsLoginResponse.TokenKey;
+                }
+                #endregion
+
+                var omsresponse = JsonConvert.DeserializeObject<UserResponse>(GetApiResponse(ConfigurationManager.AppSettings["ApiGatewayOMSURL"]
+                    + "/v1/order/updateorderstatus", Method.POST, omsRequest, token));
+                if (response != null)
+                {
+                    response.StatusMessage += ". " + omsresponse.StatusMessage;
+                }
+                #endregion
+            }
+            #endregion
+            #region Invoice Generation
+            if (request.Requests[0].OrderStatusCode == "12")
+            {
+                InvoiceResponse invoiceResponse = GetInvoiceRequest(request);
+                if (invoiceResponse != null && invoiceResponse.Data.Count > 0)
+                {
+                    InvoiceRequest invoiceRequest = new InvoiceRequest();
+                    invoiceRequest.Requests = invoiceResponse.Data;
+                    if (!String.IsNullOrEmpty(request.Token))
+                    {
+                        JsonConvert.DeserializeObject<InvoiceResponse>(GetApiResponse(ConfigurationManager.AppSettings["ApiGatewayTMSURL"] + "v1/invoice/generateinvoice", Method.POST, invoiceRequest, request.Token));
+                    }
+                }
+            }
+            #endregion
+
+            return response;
         }
 
         public override OrderStatusResponse CancelOrder(OrderStatusRequest request)
         {
-            return _orderRepository.CancelOrder(request);
+            OrderStatusResponse response = _orderRepository.CancelOrder(request);
+
+            if (response.StatusCode == (int)HttpStatusCode.OK && response.Status == DomainObjects.Resource.ResourceData.Success)
+            {
+                #region Update Status to OMS 
+                OrderStatusRequest omsRequest = new OrderStatusRequest()
+                {
+                    Requests = new List<OrderStatus>()
+                };
+
+                foreach (var item in request.Requests)
+                {
+                    OrderStatus requestData = new OrderStatus()
+                    {
+                        IsLoad = null,
+                        OrderNumber = item.OrderNumber,
+                        OrderStatusCode = item.OrderStatusCode,
+                        Remarks = "",
+                        SequenceNumber = item.SequenceNumber,
+                        NewSequenceNumber = item.NewSequenceNumber
+                    };
+                    omsRequest.Requests.Add(requestData);
+                    omsRequest.RequestFrom = "TMS";
+                }
+                if (omsRequest.Requests.Count > 0)
+                {
+                    #region Call OMS API to Update Order
+                    #region Login to OMS and get Token
+                    LoginRequest loginRequest = new LoginRequest();
+                    string token = "";
+
+                    loginRequest.UserName = ConfigurationManager.AppSettings["OMSLogin"];
+                    loginRequest.UserPassword = ConfigurationManager.AppSettings["OMSPassword"];
+                    var tmsLoginResponse = JsonConvert.DeserializeObject<UserResponse>(GetApiResponse(ConfigurationManager.AppSettings["ApiGatewayOMSURL"]
+                        + "/v1/user/login", Method.POST, loginRequest, null));
+                    if (tmsLoginResponse != null && tmsLoginResponse.Data.Count > 0)
+                    {
+                        token = tmsLoginResponse.TokenKey;
+                    }
+                    #endregion
+
+                    var omsresponse = JsonConvert.DeserializeObject<UserResponse>(GetApiResponse(ConfigurationManager.AppSettings["ApiGatewayOMSURL"]
+                        + "/v1/order/cancelorder", Method.POST, omsRequest, token));
+                    response.StatusMessage += ". " + omsresponse.StatusMessage;
+                    #endregion
+                }
+                #endregion
+
+                #region Update Status to DMS 
+                OrderStatusRequest dmsRequest = new OrderStatusRequest()
+                {
+                    Requests = new List<OrderStatus>()
+                };
+
+                foreach (var item in request.Requests)
+                {
+                    OrderStatus requestData = new OrderStatus()
+                    {
+                        IsLoad = null,
+                        OrderNumber = item.OrderNumber,
+                        OrderStatusCode = item.OrderStatusCode,
+                        Remarks = "",
+                        SequenceNumber = item.SequenceNumber,
+                        NewSequenceNumber = item.NewSequenceNumber
+                    };
+                    dmsRequest.Requests.Add(requestData);
+                    dmsRequest.RequestFrom = "TMS";
+                }
+                if (dmsRequest.Requests.Count > 0)
+                {
+                    #region Call DMS API to Update Order
+                    #region Login to DMS and get Token
+                    LoginRequest dmsLoginRequest = new LoginRequest();
+                    string dmsToken = "";
+
+                    dmsLoginRequest.UserName = ConfigurationManager.AppSettings["DMSLogin"];
+                    dmsLoginRequest.UserPassword = ConfigurationManager.AppSettings["DMSPassword"];
+                    var dmsLoginResponse = JsonConvert.DeserializeObject<UserResponse>(GetApiResponse(ConfigurationManager.AppSettings["ApiGatewayDMSURL"]
+                        + "/v1/user/login", Method.POST, dmsLoginRequest, null));
+                    if (dmsLoginResponse != null && dmsLoginResponse.Data.Count > 0)
+                    {
+                        dmsToken = dmsLoginResponse.TokenKey;
+                    }
+                    #endregion
+
+                    var dmsresponse = JsonConvert.DeserializeObject<UserResponse>(GetApiResponse(ConfigurationManager.AppSettings["ApiGatewayDMSURL"]
+                        + "/v1/trip/cancelorder", Method.POST, omsRequest, dmsToken));
+                    response.StatusMessage += ". " + dmsresponse.StatusMessage;
+                    #endregion
+                }
+                #endregion
+            }
+            return response;
         }
 
         public override HargaResponse GetHarga(HargaRequest request)
@@ -814,6 +1027,178 @@ namespace TMS.BusinessGateway.Task
 
             return orderStatusResponse;
         }
-    }
 
+        public override ImageGuidsResponse GetShippingListGuids(string orderNumber)
+        {
+            #region Login to DMS and get Token
+            LoginRequest loginRequest = new LoginRequest();
+            string token = "";
+
+            loginRequest.UserName = ConfigurationManager.AppSettings["DMSLogin"];
+            loginRequest.UserPassword = ConfigurationManager.AppSettings["DMSPassword"];
+            var dmsLoginResponse = JsonConvert.DeserializeObject<UserResponse>(GetApiResponse(ConfigurationManager.AppSettings["ApiGatewayDMSURL"]
+                + "/v1/user/login", Method.POST, loginRequest, null));
+            if (dmsLoginResponse != null && dmsLoginResponse.Data.Count > 0)
+            {
+                token = dmsLoginResponse.TokenKey;
+            }
+            #endregion
+
+            ImageGuidsResponse response = JsonConvert.DeserializeObject<ImageGuidsResponse>(GetApiResponse(ConfigurationManager.AppSettings["ApiGatewayDMSURL"]
+                + "/v1/trip/getshippinglistguids?orderNumber=" + orderNumber, Method.GET, null, token));
+
+            return response;
+        }
+
+        public override ImageGuidsResponse GetPodGuids(string orderNumber)
+        {
+            #region Login to DMS and get Token
+            LoginRequest loginRequest = new LoginRequest();
+            string token = "";
+
+            loginRequest.UserName = ConfigurationManager.AppSettings["DMSLogin"];
+            loginRequest.UserPassword = ConfigurationManager.AppSettings["DMSPassword"];
+            var dmsLoginResponse = JsonConvert.DeserializeObject<UserResponse>(GetApiResponse(ConfigurationManager.AppSettings["ApiGatewayDMSURL"]
+                + "/v1/user/login", Method.POST, loginRequest, null));
+            if (dmsLoginResponse != null && dmsLoginResponse.Data.Count > 0)
+            {
+                token = dmsLoginResponse.TokenKey;
+            }
+            #endregion
+
+            ImageGuidsResponse response = JsonConvert.DeserializeObject<ImageGuidsResponse>(GetApiResponse(ConfigurationManager.AppSettings["ApiGatewayDMSURL"]
+                + "/v1/trip/getpodguids?orderNumber=" + orderNumber, Method.GET, null, token));
+
+            return response;
+        }
+
+        public override ImageGuidsResponse GetPhotoWithCustomerGuids(string orderNumber)
+        {
+            #region Login to DMS and get Token
+            LoginRequest loginRequest = new LoginRequest();
+            string token = "";
+
+            loginRequest.UserName = ConfigurationManager.AppSettings["DMSLogin"];
+            loginRequest.UserPassword = ConfigurationManager.AppSettings["DMSPassword"];
+            var dmsLoginResponse = JsonConvert.DeserializeObject<UserResponse>(GetApiResponse(ConfigurationManager.AppSettings["ApiGatewayDMSURL"]
+                + "/v1/user/login", Method.POST, loginRequest, null));
+            if (dmsLoginResponse != null && dmsLoginResponse.Data.Count > 0)
+            {
+                token = dmsLoginResponse.TokenKey;
+            }
+            #endregion
+
+            ImageGuidsResponse response = JsonConvert.DeserializeObject<ImageGuidsResponse>(GetApiResponse(ConfigurationManager.AppSettings["ApiGatewayDMSURL"]
+                + "/v1/trip/getphotowithcustomerguids?orderNumber=" + orderNumber, Method.GET, null, token));
+
+            return response;
+        }
+
+        public override OrderResponse GetShipmentSchedulesFromEmail(string token)
+        {
+            OrderResponse tmsOrderResponse = new OrderResponse();
+
+            // Loin into Gamil and Get all Mails
+            var mailRepository = new MailRepository(
+                                    ConfigurationManager.AppSettings["GmailURL"],
+                                    Convert.ToInt32(ConfigurationManager.AppSettings["GmailURLPort"].ToString()),
+                                    true,
+                                    ConfigurationManager.AppSettings["UserEmailID"],
+                                    ConfigurationManager.AppSettings["UserPassword"]
+                                );
+
+            //Get all Unread Mails
+            var emailList = mailRepository.GetUnreadMails("inbox");
+
+            foreach (ActiveUp.Net.Mail.Message email in emailList)
+            {
+                if (email.Attachments.Count > 0)
+                {
+                    foreach (MimePart attachment in email.Attachments)
+                    {
+                        if (attachment.ContentType.MimeType.ToLower() == "application/pdf") // Checking For PDF files
+                        {
+                            // Uploading File into Blob storage and get GUID
+                            var fileuploadresponse = JsonConvert.DeserializeObject<ResponseDataForFileUpload>(GetFileUploadApiResponse(ConfigurationManager.AppSettings["ApiGatewayTMSURL"] + "/v1/media/uploadfile", Method.POST, attachment, null));
+
+                            if (fileuploadresponse.StatusCode == (int)HttpStatusCode.OK && !String.IsNullOrEmpty(fileuploadresponse.Guid))
+                            {
+                                // Calling OCR to get shipmentschedule data
+                                var res = GetFileUploadApiResponse(ConfigurationManager.AppSettings["ApiGatewayTMSURL"] + "/v1/order/shipmentscheduleocr", Method.POST, attachment, null);
+                                var shipmentScheduleOcr = JsonConvert.DeserializeObject<dynamic>(res);
+                                var jsonObject = JObject.Parse(shipmentScheduleOcr);
+                                ShipmentScheduleOcrRequest shipmentScheduleOcrRequest = new ShipmentScheduleOcrRequest();
+                                if (jsonObject.success)
+                                {
+                                    shipmentScheduleOcrRequest.Requests = new List<ShipmentScheduleOcr>();
+                                    ShipmentScheduleOcr shipmentSchedule = new ShipmentScheduleOcr()
+                                    {
+                                        ImageGUID = fileuploadresponse.Guid,
+                                        DocumentType = jsonObject.documentType,
+                                        Success = jsonObject.success,
+                                        Image = jsonObject.image,
+                                        EmailFrom = email.From.Email,
+                                        EmailDateTime = email.Date,
+                                        EmailSubject = email.Subject,
+                                        EmailText = email.BodyHtml.TextStripped,
+                                        Data = new DetailsOcr()
+                                        {
+                                            DayShipment = jsonObject.data.DayShipment,
+                                            EstimatedTotalPallet = jsonObject.data.EstimatedTotalPallet,
+                                            MainDealerCode = jsonObject.data.MainDealerCode,
+                                            MainDealerName = jsonObject.data.MainDealerName,
+                                            MultiDropShipment = jsonObject.data.MultiDropShipment,
+                                            ShipmentScheduleNo = jsonObject.data.ShipmentScheduleNo,
+                                            ShipmentTime = jsonObject.data.ShipmentTime,
+                                            ShipToParty = jsonObject.data.ShipToParty,
+                                            VehicleType = jsonObject.data.VehicleType,
+                                            Weight = jsonObject.data.Weight
+                                        }
+                                    };
+                                    shipmentScheduleOcrRequest.Requests.Add(shipmentSchedule);
+                                }
+
+                                if (shipmentScheduleOcrRequest.Requests.Count > 0)
+                                {
+                                    //OcrOrderResponse
+                                    OrderRequest omsOrderRequest = new OrderRequest();
+                                    OrderResponse ocrOrderResponse = OcrOrderResponse(shipmentScheduleOcrRequest);
+                                    if (ocrOrderResponse != null && ocrOrderResponse.StatusCode == (int)HttpStatusCode.OK)
+                                    {
+                                        LoginRequest omsLoginRequest = new LoginRequest();
+                                        string omsToken = "";
+                                        omsLoginRequest.UserName = ConfigurationManager.AppSettings["OMSLogin"];
+                                        omsLoginRequest.UserPassword = ConfigurationManager.AppSettings["OMSPassword"];
+                                        var tmsLoginResponse = JsonConvert.DeserializeObject<UserResponse>(GetApiResponse(ConfigurationManager.AppSettings["ApiGatewayOMSURL"]
+                                                                                                                          + "v1/user/login", Method.POST, omsLoginRequest, null));
+                                        if (tmsLoginResponse != null && tmsLoginResponse.Data.Count > 0)
+                                        {
+                                            omsToken = tmsLoginResponse.TokenKey;
+                                        }
+                                        omsOrderRequest.Requests = ocrOrderResponse.Data;
+
+                                        OrderResponse omsOrderResponse = JsonConvert.DeserializeObject<OrderResponse>(GetApiResponse(ConfigurationManager.AppSettings["ApiGatewayOMSURL"]
+                                                                                                             + "v1/order/createordersfromshipmentlistocr", Method.POST, omsOrderRequest, omsToken));
+
+                                        if (omsOrderResponse.Data != null && omsOrderResponse.StatusCode == (int)HttpStatusCode.OK)
+                                        {
+                                            if (!String.IsNullOrEmpty(token))
+                                            {
+                                                OrderRequest tmsOrderRequest = new OrderRequest();
+                                                tmsOrderRequest.Requests = omsOrderResponse.Data;
+                                                tmsOrderRequest.UploadType = 3;
+                                                tmsOrderResponse = JsonConvert.DeserializeObject<OrderResponse>(GetApiResponse(ConfigurationManager.AppSettings["ApiGatewayTMSURL"]
+                                                                                                                                                           + "v1/order/createordersfromshipmentlistocr", Method.POST, tmsOrderRequest, token));
+                                            }
+                                        }
+                                    }
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+            return tmsOrderResponse;
+        }
+    }
 }
