+++ TMS.BusinessGateway/packages.config
@@ -1,5 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="MailSystem.Net-trunk" version="2.1.41" targetFramework="net461" />
   <package id="Newtonsoft.Json" version="12.0.2" targetFramework="net461" />
   <package id="RestSharp" version="106.6.10" targetFramework="net461" />
 </packages>
