+++ TMS.DomainGateway/Task/Order.cs
@@ -31,5 +31,9 @@ namespace TMS.DomainGateway.Task
         public abstract OrderResponse CreateOrdersFromShipmentListOCR(OrderRequest request);
         public abstract InvoiceResponse GetInvoiceRequest(OrderStatusRequest request);
         public abstract OrderStatusResponse SwapeStopPoints(OrderStatusRequest orderStatusRequest);
+        public abstract ImageGuidsResponse GetShippingListGuids(string orderNumber);
+        public abstract ImageGuidsResponse GetPodGuids(string orderNumber);
+        public abstract ImageGuidsResponse GetPhotoWithCustomerGuids(string orderNumber);
+        public abstract OrderResponse GetShipmentSchedulesFromEmail(string token);
     }
 }
